lightweight-image-classifier/requirements.txt
# ============================================================
# PyTorch Lightweight Image Classifier - Requirements
# ============================================================

# Core Deep Learning Framework
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# Data Processing & Augmentation
numpy>=1.21.0
Pillow>=9.0.0
opencv-python>=4.6.0
albumentations>=1.3.0

# Training Utilities
tqdm>=4.64.0
pyyaml>=6.0

# Visualization & Logging
tensorboard>=2.10.0
matplotlib>=3.5.0
seaborn>=0.12.0

# Model Export
onnx>=1.13.0
onnxruntime>=1.13.0

# Code Quality
flake8>=5.0.0
black>=22.0.0
isort>=5.10.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Optional: Jupyter Notebook Support
# jupyter>=1.0.0
# ipywidgets>=8.0.0